+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/TransactionFactoryTest.java
@@ -51,7 +51,7 @@ public class TransactionFactoryTest {
     }
 
     @Test
-    public  void testCanTransferSchedulWithMemoFactory() {
+    public void testCanTransferSchedulWithMemoFactory() {
         val builder = TransactionFactory.newScheduledTransferWithMemo();
         assertNotNull(builder);
     }
@@ -65,6 +65,44 @@ public class TransactionFactoryTest {
     @Test
     public void testCanTransferToEncryptedFactory() {
         val builder = TransactionFactory.newTransferToEncrypted();
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testCanRemoveBakerFactory() {
+        val builder = TransactionFactory.newRemoveBaker();
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testCanConfigureBakerFactory() {
+        val builder = TransactionFactory.newConfigureBaker();
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testCanUpdateBakerStakeFactory() {
+        val stake = CCDAmount.fromMicro("10000000");
+        val builder = TransactionFactory.newUpdateBakerStake(stake);
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testCanUpdateBakerRestakeEarningsFactory() {
+        val builder = TransactionFactory.newUpdateBakerRestakeEarnings(true);
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testCanUpdateBakerKeysFactory() {
+        AccountAddress sender = AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc");
+        val builder = TransactionFactory.newUpdateBakerKeys(sender);
+        assertNotNull(builder);
+    }
+
+    @Test
+    public void testConfigureDelegationFactory() {
+        val builder = TransactionFactory.newConfigureDelegation();
         assertNotNull(builder);
     }
 }