+++ crypto-jni/src/lib.rs
@@ -1,6 +1,7 @@
 use concordium_base::common::derive::Serialize;
 pub use concordium_base::common::types::{AccountAddress, ACCOUNT_ADDRESS_SIZE};
-use concordium_base::common::*;
+use concordium_base::transactions::{BakerKeysPayload, AddBakerKeysMarker, ConfigureBakerKeysPayload};
+use concordium_base::{common::*, base};
 use concordium_base::contracts_common::Amount;
 use concordium_base::encrypted_transfers;
 use concordium_base::encrypted_transfers::types::{
@@ -357,4 +358,51 @@ pub extern "system" fn Java_com_concordium_sdk_crypto_CryptoJniNative_generateEn
         Some(payload) => CryptoJniResult::Ok(payload).to_jstring(&env),
         None => EncryptedAmountTransferResult::Err(PAYLOAD_CREATION_ERROR).to_jstring(&env),
     }
+}
+
+#[no_mangle]
+/// The JNI wrapper for the `generate_baker_keys` method.
+pub extern "system" fn Java_com_concordium_sdk_crypto_CryptoJniNative_generateBakerKeys(
+    env: JNIEnv,
+    _: JClass,
+) -> jstring {
+    let mut csprng = thread_rng();
+    let payload = base::BakerKeyPairs::generate(&mut csprng);
+    return CryptoJniResult::Ok(payload).to_jstring(&env);
+}
+
+#[derive(Serialize, SerdeSerialize, SerdeDeserialize)]
+#[serde(rename_all = "camelCase")]
+struct AddBakerPayloadInput {
+    pub sender:   AccountAddress,
+    pub keys:     base::BakerKeyPairs,
+}
+
+type AddBakerResult = CryptoJniResult<BakerKeysPayload<AddBakerKeysMarker>>;
+
+#[no_mangle]
+#[allow(non_snake_case)]
+/// The JNI wrapper to create new configure baker keys payload.
+/// Construct a BakerKeysPayload with proofs for updating baker keys.
+pub extern "system" fn Java_com_concordium_sdk_crypto_CryptoJniNative_generateConfigureBakerKeysPayload(
+    env: JNIEnv,
+    _: JClass,
+    input: JString,
+) -> jstring {
+    let input: AddBakerPayloadInput = match env.get_string(input) {
+        Ok(java_str) => match java_str.to_str() {
+            Ok(rust_str) => match from_str(rust_str) {
+                Ok(input) => input,
+                Err(err) => return AddBakerResult::from(err).to_jstring(&env),
+            },
+            Err(err) => return AddBakerResult::from(err).to_jstring(&env),
+        },
+        Err(err) => return AddBakerResult::from(err).to_jstring(&env),
+    };
+
+    let mut csprng = thread_rng();
+
+    let payload = ConfigureBakerKeysPayload::new(&input.keys, input.sender, &mut csprng);
+
+    return CryptoJniResult::Ok(payload).to_jstring(&env);
 }